change the struct of the project
